+++ tests/Builders/TokenPreauthBuilder.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Tests\Builders;
+
+class TokenPreauthBuilder extends TokenPaymentBuilder
+{
+}
